+++ example/CHANGELOG.md
@@ -1,5 +1,12 @@
 # example
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies
+  - cesium-wind-layer@0.4.1
+
 ## 0.3.0
 
 ### Minor Changes